# openapi.yaml
openapi: 3.0.3
info:
  title: School Reading & Assessment Platform API
  version: 1.0.0
  description: |
    Complete REST API for the school oral reading and AI assessment platform.

    ## Authentication
    All endpoints except `/auth/register` and `/auth/login` require a JWT Bearer token
    in the `Authorization: Bearer <token>` header. Access tokens expire in 15 minutes.
    Use `/auth/refresh` to obtain a new one without re-logging in.

    ## Roles
    | Role    | Access                                                        |
    |---------|---------------------------------------------------------------|
    | admin   | School structure, user management, all assignments            |
    | teacher | Their assigned sections, lesson upload, student progress view |
    | student | Their section's lessons, reading sessions, personal dashboard |

    ## Async Assessment Flow
    1. `POST /student/sessions` — creates session record
    2. Student records audio in browser via MediaRecorder API
    3. `POST /student/sessions/{id}/submit` — uploads audio (WebM/OGG)
    4. Server streams audio to Whisper STT microservice (audio never stored)
    5. Transcript sent to Ollama LLM for assessment (BullMQ job)
    6. **Socket.io** emits `assessment.complete` to student's browser
    7. `GET /student/sessions/{id}/result` — fetch the assessment result

    ## Socket.io Events
    Connect with the JWT token as a query param: `ws://server/api/v1?token=<jwt>`

...
servers:
...
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /auth/login

  # ── Reusable Parameters ───────────────────────────────────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Results per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    SectionIdParam:
      name: sectionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    LessonIdParam:
      name: lessonId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2026-05-30T08:30:00Z"
    Language:
      type: string
      enum: [english, tamil]
      example: english
...
    SessionStatus:
      type: string
      enum: [in_progress, completed, abandoned]
      example: completed
    AssessmentStatus:
      type: string
      enum: [pending, processing, completed, failed]
      example: completed
    PdfExtractionStatus:
      type: string
      enum: [pending, success, failed]
      example: success
    CompletionStatus:
      type: string
      enum: [not_started, in_progress, completed]
      example: completed

    # ── Shared Structures ─────────────────────────────────────────────────
    Pagination:
      type: object
      required: [total, page, limit, total_pages]
      properties:
        total:
          type: integer
          example: 150
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 8
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
          example: Resource not found
        details:
          type: object
          additionalProperties: true
          nullable: true
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: must be a valid email address
    ValidationError:
      type: object
      required: [code, message, errors]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request validation failed
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'

    # ── Auth Schemas ──────────────────────────────────────────────────────
...
    StudentAssignment:
      type: object
      required: [id, student_id, student_name, student_email, section_id, section_name, class_name, is_active, assigned_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        student_id:
          $ref: '#/components/schemas/UUID'
        student_name:
          type: string
          example: Arun Prakash
        student_email:
          type: string
          example: arun@school.edu
        section_id:
          $ref: '#/components/schemas/UUID'
        section_name:
          type: string
          example: Section B
        class_name:
          type: string
          example: Grade 5
        assigned_by_name:
          type: string
          example: Admin
        is_active:
          type: boolean
          example: true
        assigned_at:
          $ref: '#/components/schemas/Timestamp'
        unassigned_at:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Timestamp'
    CreateStudentAssignmentRequest:
      type: object
      required: [student_id, section_id]
      properties:
        student_id:
          $ref: '#/components/schemas/UUID'
        section_id:
          $ref: '#/components/schemas/UUID'
...
    Lesson:
      type: object
      required: [id, section_id, uploaded_by_id, uploaded_by_name, title, language,
                 pdf_extraction_status, display_order, is_published, created_at, updated_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        section_id:
          $ref: '#/components/schemas/UUID'
        uploaded_by_id:
          $ref: '#/components/schemas/UUID'
        uploaded_by_name:
          type: string
          example: Mrs. Anitha
        title:
          type: string
          example: Chapter 3 - The Water Cycle
        description:
          type: string
          nullable: true
          example: Reading comprehension on the water cycle
        language:
          $ref: '#/components/schemas/Language'
        pdf_extraction_status:
          $ref: '#/components/schemas/PdfExtractionStatus'
        display_order:
          type: integer
          example: 3
        is_published:
          type: boolean
          example: true
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
    LessonWithStats:
      allOf:
        - $ref: '#/components/schemas/Lesson'
        - type: object
          properties:
            attempt_count:
              type: integer
              description: Total reading sessions by all students
              example: 47
            avg_score:
              type: number
              format: float
              nullable: true
              description: Average overall_score across all student sessions
              example: 76.3
    LessonWithStudentProgress:
      allOf:
        - $ref: '#/components/schemas/Lesson'
        - type: object
          required: [my_best_score, my_attempt_count, completion_status]
          properties:
            my_best_score:
              type: number
              format: float
              nullable: true
              description: Student's highest overall_score across all attempts
              example: 85.5
            my_attempt_count:
              type: integer
              example: 2
            completion_status:
              $ref: '#/components/schemas/CompletionStatus'
    UpdateLessonRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          example: Chapter 3 - The Water Cycle (Revised)
        description:
          type: string
          nullable: true
        language:
          $ref: '#/components/schemas/Language'
        is_published:
          type: boolean
          example: true
...
    PdfUrlResponse:
      type: object
      required: [url, expires_at]
      properties:
        url:
          type: string
          description: MinIO presigned URL valid for 15 minutes
          example: http://school-server/storage/lessons/uuid.pdf?X-Amz-Signature=abc123
        expires_at:
          $ref: '#/components/schemas/Timestamp'

    # ── Reading Session Schemas ────────────────────────────────────────────
    ReadingSession:
      type: object
      required: [id, student_id, lesson_id, lesson_title, attempt_number,
                 status, language, assessment_status, started_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        student_id:
          $ref: '#/components/schemas/UUID'
        lesson_id:
          $ref: '#/components/schemas/UUID'
        lesson_title:
          type: string
          example: Chapter 3 - The Water Cycle
        attempt_number:
          type: integer
          example: 2
        status:
          $ref: '#/components/schemas/SessionStatus'
        language:
          $ref: '#/components/schemas/Language'
        assessment_status:
          $ref: '#/components/schemas/AssessmentStatus'
        duration_seconds:
          type: integer
          nullable: true
          example: 187
        started_at:
          $ref: '#/components/schemas/Timestamp'
        completed_at:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Timestamp'
...
    StudentProgressRow:
      type: object
      required: [student_id, student_name, lessons_completed, total_sessions, current_streak]
      properties:
        student_id:
          $ref: '#/components/schemas/UUID'
        student_name:
          type: string
          example: Arun Prakash
        lessons_completed:
          type: integer
          example: 6
        total_sessions:
          type: integer
          example: 14
        avg_overall_score:
          type: number
          format: float
          nullable: true
          example: 78.2
        avg_accuracy_score:
          type: number
          format: float
          nullable: true
          example: 84.2
        avg_fluency_score:
          type: number
          format: float
          nullable: true
          example: 71.5
        current_streak:
          type: integer
          example: 7
        last_session_date:
          type: string
          format: date
          nullable: true
          example: "2026-05-30"
        current_rank:
          type: integer
          nullable: true
          example: 3
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Invalid or expired token
    Forbidden:
      description: Authenticated but insufficient role permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: FORBIDDEN
            message: You do not have permission to access this resource
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: Resource not found
...
    UnprocessableEntity:
      description: Request body or query parameters failed schema validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
paths:
...
  /admin/assignments/students:
    get:
      tags: [Admin — Student Assignments]
      summary: List student-section assignments
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: section_id
          in: query
          schema:
            type: string
            format: uuid
        - name: is_active
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Paginated assignment list
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StudentAssignment'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Admin — Student Assignments]
      summary: Assign a student to a section
      description: A student can only have one active section assignment. Returns 409 if already assigned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentAssignmentRequest'
            example:
              student_id: "550e8400-e29b-41d4-a716-446655440003"
              section_id: "550e8400-e29b-41d4-a716-446655440002"
      responses:
        '201':
          description: Student assigned to section
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentAssignment'
        '400':
          description: Specified user is not a student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: INVALID_ROLE
                message: The specified user does not have the student role
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Student is already assigned to a section
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: ALREADY_ASSIGNED
                message: "Student is already assigned to Grade 5 - Section A. Use PUT /admin/assignments/students/{studentId}/move to reassign."
...
  /teacher/sections/{sectionId}/lessons:
    get:
      tags: [Teacher — Lessons]
      summary: List all lessons in a section (teacher view with stats)
      parameters:
        - $ref: '#/components/parameters/SectionIdParam'
        - name: is_published
          in: query
          description: Filter by published state
          schema:
            type: boolean
      responses:
        '200':
          description: Lessons ordered by display_order, with section-wide attempt stats
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LessonWithStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Section not found or not assigned to this teacher
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Teacher — Lessons]
      summary: Upload a new PDF lesson to a section
      description: |
        Accepts multipart/form-data. The PDF is stored in MinIO.
        Text extraction from the PDF runs asynchronously — `pdf_extraction_status`
        will be `pending` initially and update to `success` or `failed`.
        Students cannot start sessions on a lesson until extraction is `success`.
      parameters:
        - $ref: '#/components/parameters/SectionIdParam'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, language, pdf]
              properties:
                title:
                  type: string
                  maxLength: 500
                  example: Chapter 3 - The Water Cycle
                description:
                  type: string
                  maxLength: 2000
                  example: Reading comprehension exercise on the water cycle
                language:
                  $ref: '#/components/schemas/Language'
                is_published:
                  type: boolean
                  default: false
                  description: If false, students cannot see the lesson yet
                pdf:
                  type: string
                  format: binary
                  description: PDF file only; max 50 MB
      responses:
        '201':
          description: Lesson record created; PDF text extraction queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '400':
          description: File is not a PDF or exceeds 50 MB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: INVALID_FILE
                message: Only PDF files up to 50 MB are accepted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /teacher/sections/{sectionId}/lessons/{lessonId}:
    get:
      tags: [Teacher — Lessons]
      summary: Get a single lesson
      parameters:
        - $ref: '#/components/parameters/SectionIdParam'
        - $ref: '#/components/parameters/LessonIdParam'
      responses:
        '200':
          description: Lesson detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonWithStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Teacher — Lessons]
      summary: Update lesson metadata
      description: Updates title, description, language, or published state. Does not replace the PDF file.
      parameters:
        - $ref: '#/components/parameters/SectionIdParam'
        - $ref: '#/components/parameters/LessonIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLessonRequest'
      responses:
        '200':
          description: Lesson updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Teacher — Lessons]
      summary: Delete a lesson and its PDF from storage
      description: |
        Hard-deletes the lesson record and removes the PDF from MinIO.
        If completed student sessions exist, the server returns **409** and requires
        `?force=true` to acknowledge that session history will also be deleted.
      parameters:
        - $ref: '#/components/parameters/SectionIdParam'
        - $ref: '#/components/parameters/LessonIdParam'
        - name: force
          in: query
          description: Set to true to confirm deletion when student session history exists
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Lesson and PDF deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Student session history exists; pass force=true to confirm
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: HAS_STUDENT_DATA
                message: "14 student sessions exist for this lesson. Add ?force=true to delete anyway."
  /teacher/sections/{sectionId}/lessons/{lessonId}/pdf-url:
    get:
      tags: [Teacher — Lessons]
      summary: Get a presigned URL to view the lesson PDF (15-minute expiry)
      parameters:
        - $ref: '#/components/parameters/SectionIdParam'
        - $ref: '#/components/parameters/LessonIdParam'
      responses:
        '200':
          description: Presigned MinIO URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdfUrlResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════════════════════════════════════
  # TEACHER — DASHBOARD
  # ═══════════════════════════════════════════════════════════════════════════
...
  /teacher/sections/{sectionId}/students:
    get:
      tags: [Teacher — Dashboard]
      summary: Get the progress table for all students in a section
      parameters:
        - $ref: '#/components/parameters/SectionIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [name, avg_overall_score, lessons_completed, current_streak, last_session_date]
            default: avg_overall_score
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: search
          in: query
          description: Partial match on student name
          schema:
            type: string
      responses:
        '200':
          description: Student progress rows
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StudentProgressRow'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /student/lessons:
    get:
      tags: [Student — Lessons]
      summary: List all published lessons in the student's section
      parameters:
        - name: status
          in: query
          description: Filter by the student's completion status
          schema:
            $ref: '#/components/schemas/CompletionStatus'
      responses:
        '200':
          description: Published lessons with student's personal progress overlay
          content:
            application/json:
              schema:
                type: object
                required: [section_name, data]
                properties:
                  section_name:
                    type: string
                    example: Grade 5 - Section A
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LessonWithStudentProgress'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '412':
          description: Student has not been assigned to a section yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: NOT_ASSIGNED
                message: You have not been assigned to a section yet. Contact your admin.
  /student/lessons/{lessonId}:
    get:
      tags: [Student — Lessons]
      summary: Get a single lesson with the student's attempt history
      parameters:
        - $ref: '#/components/parameters/LessonIdParam'
      responses:
        '200':
          description: Lesson with past attempts
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/LessonWithStudentProgress'
                  - type: object
                    properties:
                      attempts:
                        type: array
                        items:
                          type: object
                          required: [session_id, attempt_number, overall_score, completed_at]
                          properties:
                            session_id:
                              $ref: '#/components/schemas/UUID'
                            attempt_number:
                              type: integer
                              example: 2
                            overall_score:
                              type: number
                              format: float
                              nullable: true
                              example: 78.2
                            completed_at:
                              $ref: '#/components/schemas/Timestamp'
                              nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /student/lessons/{lessonId}/pdf-url:
    get:
      tags: [Student — Lessons]
      summary: Get a presigned URL to view the lesson PDF (15-minute expiry)
      description: Pass this URL to react-pdf in the browser. The PDF is never downloaded — it is rendered in-browser only.
      parameters:
        - $ref: '#/components/parameters/LessonIdParam'
      responses:
        '200':
          description: Presigned MinIO URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdfUrlResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: PDF text extraction failed; lesson cannot be viewed until teacher re-uploads
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: PDF_EXTRACTION_FAILED
                message: This lesson could not be processed. Your teacher has been notified.
  /student/lessons/{lessonId}/sessions:
    get:
      tags: [Student — Sessions]
      summary: List all of the student's past reading sessions for a lesson
      parameters:
        - $ref: '#/components/parameters/LessonIdParam'
      responses:
        '200':
          description: All attempts for this lesson, newest first
          content:
            application/json:
              schema:
                type: object
                required: [lesson_id, lesson_title, data]
                properties:
                  lesson_id:
                    $ref: '#/components/schemas/UUID'
                  lesson_title:
                    type: string
                    example: Chapter 3 - The Water Cycle
                  data:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/ReadingSession'
                        - type: object
                          properties:
                            overall_score:
                              type: number
                              format: float
                              nullable: true
                              example: 78.2
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════════════════════════════════════
  # STUDENT — SESSIONS
  # ═══════════════════════════════════════════════════════════════════════════
...